// frontend/src/app/model/request.ts
export class request { 
  _id:number;
  documentNames:[];
  vendorOrg:String;
  products:[];
  market: String;
  date: String;
  criticality:String;
  comment: String;
  allComments: [];
  respondentName: String;
  respondentEmail: String;
  respondentContact: String;
  uploadDate: String;
  requestStatus: String;
}